fix: undo unnecessary changes
